Default new zones to a blank root password.
If no password has been set, you'd have to log in with zlogin -S
This is obtuse and highly undesireable.  A blank password is much
better as it can only be used to zlogin the first time anyway
and if someone were to turn on ssh (with PermitRootLogin yes)
it would still prevent a remote login.