t_printengine & friends: avoid bullshit infinite loop
What's happening is that when the slave side of the pty is closed, reads
on the master side are never failing (or, in fact, returning), we loop
forever.

As a workaround, we check for short reads (anything less than the 1024 we feed
into the read).  If it's short, we write what's left then terminate, assuming
that something at the 0-length read we need to indicate EOF.