Include subjectAltName in CA certificates so that it is there for when it signs certificates which set issuerAltName=issuer:copy. Including the email address in the subject alt name of a CA is a PKIX recommendation anyway.
issuer:copy means to copy the subjectAltNames from the issuing certificate
but if the issuer has none then OpenSSL creates a certificate with an empty
issuerAltName section:

            X509v3 Issuer Alternative Name:
                <EMPTY>

which is not RFC compliant.

The cryptography python module enforces RFC compliance for this field as of
version 2.2 and as a result our unsupported critical extension tests fail.