cloned zones do not pick up IP information from zone configuration.
An OmniOS user reported that a cloned zone would not pick up IP information
from the zone configuration until the appropriate IP interface was manually
configured and then the zone rebooted.

This turned out to be due to the ipmgmtd/first_boot_done property on the
network/ip-interface-management:default service not being unset by the
system-unconfigure script which is responsible for unconfiguring a zone
following cloning.