Only check minimum permissions on catalogue files
If the catalogue is on a filesystem which has additional ACLs,
for example one which enforces group writeability, which is
common to allow all members of a group to manipulate the
repository, we currently fail because the mode of the catalogue
files is not exactly what we expect (0644).

This change allows the actual permissions to be a superset of
the minimum requirement.